style: Restylization of code in accordance with standardized practices of atomization.
- Moved `CompendiumTopAppBar` to its own composable and placed it in the utils app-wide package because it is used throughout the app.
- Moved `InputSearchPhrase` to global utils similar to `CompendiumTopAppBar`, as it is a helper function used within it.
- Moved `CompendiumUiStateComposable` and its associated states into a separate file.
- Moved `CompendiumItemScreen` and its sub-functions to their own separate files.
- Separated ui files into packages: `ItemsGrid` and `SelectedItem`
