+++ app/src/main/java/luke/koz/hyrulecompendium/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import luke.koz.hyrulecompendium.ui.screens.CompendiumAppScreen
+import luke.koz.hyrulecompendium.ui.screens.ItemsGrid.CompendiumAppScreen
 import luke.koz.hyrulecompendium.ui.theme.HyruleCompendiumTheme
 
 class MainActivity : ComponentActivity() {
