+++ app/src/main/java/luke/koz/hyrulecompendium/ui/screens/ItemsGrid/ErrorScreen.kt
@@ -1,0 +1,47 @@
+package luke.koz.hyrulecompendium.ui.screens.ItemsGrid
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import luke.koz.hyrulecompendium.R
+import luke.koz.hyrulecompendium.ui.theme.HyruleCompendiumTheme
+
+
+/**
+ * The home screen displaying error message with re-attempt button.
+ */
+@Composable
+fun ErrorScreen(retryAction: () -> Unit, modifier: Modifier = Modifier) {
+    Column(
+        modifier = modifier,
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.ic_connection_error), contentDescription = ""
+        )
+        Text(text = stringResource(R.string.loading_failed), modifier = Modifier.padding(16.dp))
+        Button(onClick = retryAction) {
+            Text(stringResource(R.string.retry))
+        }
+    }
+}
+
+
+@Preview(showBackground = true)
+@Composable
+fun ErrorScreenPreview() {
+    HyruleCompendiumTheme {
+        ErrorScreen({})
+    }
+}
